workshops/Protein_Language_Modelling/scripts/training/esm-accelerate-examples/requirements.txt
transformers
datasets
accelerate
torchinfo
bitsandbytes
nvidia-ml-py3
peft==0.4.0